requirements.txt
Flask==1.0.2
flask-wtf==0.14.2
python-docx==0.8.7
PyPDF2==1.26.0
num2words==0.5.7

